Output fix.

+++ src/PulsarSearch.cpp
@@ -250,7 +250,7 @@ int main(int argc, char * argv[]) {
   cl::NDRange transposeGlobal(obs.getNrSamplesPerPaddedSecond(), static_cast< unsigned int >(std::ceil(static_cast< double >(obs.getNrDMs()) / transposeParameters[deviceName][obs.getNrDMs()])));
   cl::NDRange transposeLocal(transposeParameters[deviceName][obs.getNrDMs()], 1);
   if ( DEBUG && world.rank() == 0 ) {
-    std::cout << "Transpose global: " << obs.getNrSamplesPerPaddedSecond() << ", " << std::ceil(static_cast< double >(obs.getNrDMs()) / transposeParameters[deviceName][obs.getNrDMs()]) << std::endl;
+    std::cout << "Transpose global: " << obs.getNrSamplesPerPaddedSecond() << ", " << static_cast< unsigned int >(std::ceil(static_cast< double >(obs.getNrDMs()) / transposeParameters[deviceName][obs.getNrDMs()])) << std::endl;
     std::cout << "Transpose local: " << transposeParameters[deviceName][obs.getNrDMs()] << ", 1" << std::endl;
   }
   cl::NDRange foldingGlobal(obs.getNrPaddedDMs() / foldingParameters[deviceName][obs.getNrDMs()][obs.getNrPeriods()][6] / foldingParameters[deviceName][obs.getNrDMs()][obs.getNrPeriods()][3], obs.getNrPeriods() / foldingParameters[deviceName][obs.getNrDMs()][obs.getNrPeriods()][6] / foldingParameters[deviceName][obs.getNrDMs()][obs.getNrPeriods()][4], obs.getNrBins() / foldingParameters[deviceName][obs.getNrDMs()][obs.getNrPeriods()][6] / foldingParameters[deviceName][obs.getNrDMs()][obs.getNrPeriods()][5]);
